Add AugurSecrets accessory and Overshields

+++ Projectiles/AcceltraProj.cs
@@ -41,7 +41,7 @@ namespace wfMod.Projectiles
                 int num353 = Dust.NewDust(projectile.position, projectile.width, projectile.height, DustID.WhiteTorch);
                 Dust dust = Main.dust[num353];
                 dust.noLight = true;
-                dust.scale = (float)Main.rand.Next(80, 130) * 0.01f;
+                dust.scale = Main.rand.Next(80, 130) * 0.01f;
                 dust.velocity *= 0.2f;
                 dust.noGravity = true;
             }
